Add files via upload
Filter the dataset
Your code to filter the dataset works well. You might want to add some checks to make sure the 'SEASON_TYPE' column is present.

Find the player with the most regular seasons
The code to find this player is correct, but consider adding a check to make sure the 'PLAYER_NAME' column exists.

Calculate the player's three-point accuracy each season
The code for calculating three-point accuracy works fine. However, it would help to add a check to make sure the 'FG3M' and 'FG3A' columns are there.

Perform linear regression on three-point accuracy over the years
The regression code is correct, but add a check to ensure the 'SEASON' column is present.

Calculate the average three-point accuracy
The calculation is fine, but again, make sure the 'SEASON' column exists by adding an error check.

Compare the average accuracy to the actual number of 3-pointers made
The comparison code works, but add a check to ensure the 'FG3M' column exists.

Interpolate missing values for 2002-2003 and 2015-2016 seasons
The interpolation code is correct, but include a check to make sure the 'SEASON' column is present.